[FEAT] 23/07/11 은행 새해선물 관리 프로그램 수정
1. Member클래스 money 데이터 타입 변경
 - int -> long
2. MemberFunction 고객정보 문자타입 변경
 - next() -> nextLine()
3. MemberRun 클래스 메소드명 변경
 - DeleteInfo -> removeInfo

+++ src/com/kh/BankGiftProgram/Member.java
@@ -8,11 +8,11 @@ public class Member {
 	private String position;
 	private String phone;
 	private String address;
-	private int money;
+	private long money;
 	
 	// ===========================> 생성자
 	public Member() {}
-	public Member(String name, String company, String position, String phone, String address, int money) {
+	public Member(String name, String company, String position, String phone, String address, long money) {
 		this.name = name;
 		this.company = company;
 		this.position = position;
@@ -44,7 +44,7 @@ public class Member {
 		return this.address;
 	}
 	
-	public int getMoney() {
+	public long getMoney() {
 		return this.money;
 	}
 	
@@ -69,7 +69,7 @@ public class Member {
 		this.address = address;
 	}
 	
-	public void setMoney(int money) {
+	public void setMoney(long money) {
 		this.money = money;
 	}
 }
